+++ lib/country_codes_plus.dart
@@ -1,3 +1,4 @@
 export 'src/country_details.dart';
+export 'src/subdivision_details.dart';
 export 'src/country_codes.dart';
 export 'src/dial_code_formatter.dart';